+++ app/controllers/api/v1/categories_controller.rb
@@ -1,4 +1,6 @@
 class Api::V1::CategoriesController < ApplicationController
+    acts_as_token_authentication_handler_for User, only: [:create, :update, :delete]
+    
     def index
         categories = Category.all
         render json: categories, status: :ok
@@ -10,23 +12,35 @@ class Api::V1::CategoriesController < ApplicationController
         render json: e, status: :bad_request
     end
     def create
-        category = Category.new(category_params)
-        category.save!
-        render json: category, status: :created
+        if current_user.is_admin
+            category = Category.new(category_params)
+            category.save!
+            render json: category, status: :created
+        else
+            render json: {message: "Você não tem permissão para isso"}, status: :ok
+        end
     rescue StandardError => e
         render json: e, status: :bad_request
     end
     def update
-        category = Category.find(params[:id])
-        category.update!(category_params)
-        render json: category, status: :ok
+        if current_user.is_admin
+            category = Category.find(params[:id])
+            category.update!(category_params)
+            render json: category, status: :ok
+        else
+            render json: {message: "Você não tem permissão para isso"}, status: :ok
+        end
     rescue StandardError => e
         render json: e, status: :bad_request
     end
     def delete
-        category = Category.find(params[:id])
-        category.destroy!
-        render json: { message: "Categoria #{category.name} destruída com sucesso" }
+        if current_user.is_admin
+            category = Category.find(params[:id])
+            category.destroy!
+            render json: { message: "Categoria #{category.name} destruída com sucesso" }
+        else
+            render json: {message: "Você não tem permissão para isso"}, status: :ok
+        end
     rescue StandardError => e
         render json: e, status: :bad_request
     end
